+++ animal3D-SDK/animal3D-SDK/resource/glsl/4x/fs/03-framebuffer/drawTexture_coordManip_fs4x.glsl
@@ -34,14 +34,14 @@
 out vec4 rtFragColor;
 
 uniform sampler2D uTex_dm;
-uniform float uTime;
+uniform double uTime;
 in vec2 vTexCoord;
 
 void main()
 {
 	// DUMMY OUTPUT: all fragments are OPAQUE DARK GREY
 	vec2 texCoord = vTexCoord;
-	texCoord += vec2(sin(uTime), 0);
+	texCoord += vec2(sin(float(uTime)), 0);
 	vec3 color = texture(uTex_dm, texCoord).xyz;
 
 	rtFragColor = vec4(color, 1.0);
